requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
gunicorn==22.0.0
django-allauth==0.63.3
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
pillow==10.3.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
